# Conflicts:
#	client/src/main/java/nl/tudelft/oopp/demo/controllers/RoomViewController.java

+++ client/src/main/java/nl/tudelft/oopp/demo/entities/Food.java
@@ -104,7 +104,7 @@ public class Food {
                 Food f = new Food();
                 f.setFoodId(jsonArrayFood.getJSONObject(i).getInt("id"));
                 f.setFoodName(jsonArrayFood.getJSONObject(i).getString("name"));
-                f.setFoodPrice(jsonArrayFood.getJSONObject(i).getInt("price"));
+                f.setFoodPrice(jsonArrayFood.getJSONObject(i).getDouble("price"));
                 foodData.add(f);
             }
             return foodData;
@@ -127,7 +127,7 @@ public class Food {
                 Food f = new Food();
                 f.setFoodId(jsonArrayFood.getJSONObject(i).getInt("id"));
                 f.setFoodName(jsonArrayFood.getJSONObject(i).getString("name"));
-                f.setFoodPrice(jsonArrayFood.getJSONObject(i).getInt("price"));
+                f.setFoodPrice(jsonArrayFood.getJSONObject(i).getDouble("price"));
                 foodData.add(f);
             }
             return foodData;
@@ -150,7 +150,7 @@ public class Food {
                 Food f = new Food();
                 f.setFoodId(jsonArrayFood.getJSONObject(i).getInt("id"));
                 f.setFoodName(jsonArrayFood.getJSONObject(i).getString("name"));
-                f.setFoodPrice(jsonArrayFood.getJSONObject(i).getInt("price"));
+                f.setFoodPrice(jsonArrayFood.getJSONObject(i).getDouble("price"));
                 foodData.add(f);
             }
             return foodData;
@@ -173,7 +173,7 @@ public class Food {
                 Food f = new Food();
                 f.setFoodId(jsonArrayFood.getJSONObject(i).getInt("id"));
                 f.setFoodName(jsonArrayFood.getJSONObject(i).getString("name"));
-                f.setFoodPrice(jsonArrayFood.getJSONObject(i).getInt("price"));
+                f.setFoodPrice(jsonArrayFood.getJSONObject(i).getDouble("price"));
                 foodData.add(f);
             }
             return foodData;
@@ -194,7 +194,7 @@ public class Food {
             Food f = new Food();
             f.setFoodId(jsonObject.getInt("id"));
             f.setFoodName(jsonObject.getString("name"));
-            f.setFoodPrice(jsonObject.getInt("price"));
+            f.setFoodPrice(jsonObject.getDouble("price"));
             return f;
         } catch (Exception e) {
             logger.log(Level.SEVERE, e.toString());
